# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Axix Finance API
  version: 0.2.0
  description: >-
    Public and admin endpoints for Axix Finance investment platform.
    This spec has been expanded with component schemas, standardized error shapes,
    and planned (future) admin observability endpoints annotated via `x-planned: true`.
servers:
  - url: https://axix-finance.fly.dev/api
    description: Production
  - url: http://localhost:8080/api
    description: Local Dev
...
security:
  - bearerAuth: []
paths:
...
  /admin/jobs/daily-investments/status:
    get:
      tags: [Jobs, Admin]
      summary: Latest and previous daily job run status
      operationId: getDailyJobStatus
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current status payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyJobStatus"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/ServerError" }
  /admin/jobs/daily-investments/runs:
    get:
      tags: [Jobs, Admin]
      summary: Paginated historical runs
      operationId: listDailyJobRuns
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 200
      responses:
        "200":
          description: Paginated run list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/JobRun"
                  meta:
                    $ref: "#/components/schemas/PaginationMeta"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/ServerError" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    JobRun:
      type: object
      properties:
        id: { type: integer }
        job_name: { type: string }
        started_at: { type: string, format: date-time }
        finished_at: { type: string, format: date-time, nullable: true }
        processed_count: { type: integer }
        completed_count: { type: integer }
        total_applied: { type: number }
        success: { type: boolean }
        error_text: { type: string, nullable: true }
        source: { type: string }
    DailyJobStatus:
      type: object
      properties:
        last: { $ref: "#/components/schemas/JobRun" }
        previous: { $ref: "#/components/schemas/JobRun" }
        stale: { type: boolean }
...
    PaginationMeta:
      type: object
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: object, nullable: true }
  responses:
    Unauthorized:
      description: Authentication required or invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            unauthorized:
              value:
                error: { code: UNAUTHORIZED, message: Authorization required }
    Forbidden:
      description: Insufficient privileges
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            forbidden:
              value:
                error: { code: FORBIDDEN, message: Admin permission required }
...
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            serverError:
              value:
                error: { code: INTERNAL_ERROR, message: Unexpected failure }