# server/admin-routes.ts
import { createClient } from "@supabase/supabase-js";
import { Router } from "express";
import { requireAdminRole } from "./auth";
import {
  sendDepositApprovedEmail,
  sendWithdrawalApprovedEmail,
} from "./emailService";
import { createHttpError } from "./wrapAsync";

const adminRouter = Router();
const supabase = createClient(
  process.env.SUPABASE_URL!,
  process.env.SUPABASE_SERVICE_ROLE_KEY!
);

// Job status / staleness endpoint
adminRouter.get(
  "/jobs/daily-investments/status",
  requireAdminRole,
  async (_req, res) => {
    try {
      const { data: runs, error } = await supabase
        .from("job_runs")
        .select("*")
        .eq("job_name", "daily-investments")
        .order("started_at", { ascending: false })
        .limit(2);
      if (error) {
        return res.status(500).json({ ok: false, error: error.message });
      }
      const last = runs?.[0] || null;
      const prev = runs?.[1] || null;
      const now = Date.now();
      const stale =
        !last ||
        now - new Date(last.started_at).getTime() > 26 * 60 * 60 * 1000;
      return res.status(200).json({ ok: true, last, previous: prev, stale });
    } catch (e: any) {
      return res.status(500).json({ ok: false, error: e.message });
    }
  }
);

// Consolidated health endpoint (superset summary) for daily job
adminRouter.get(
  "/jobs/daily-investments/health",
  requireAdminRole,
  async (_req, res, next) => {
    try {
      const { data: runs, error } = await supabase
        .from("job_runs")
        .select("*")
        .eq("job_name", "daily-investments")
        .order("started_at", { ascending: false })
        .limit(7); // last week of runs
      if (error)
        return next(
          createHttpError(
            500,
            "JOB_HEALTH_QUERY_FAIL",
            "Failed to query job runs",
            { error: error.message }
          )
        );
      const last = runs?.[0] || null;
      const nowMs = Date.now();
      const stale =
        !last ||
        nowMs - new Date(last.started_at).getTime() > 26 * 60 * 60 * 1000;
      const successes = (runs || []).filter((r) => r.success).length;
      const failures = (runs || []).filter((r) => !r.success).length;
      const avgProcessed =
        runs && runs.length
          ? Math.round(
              runs.reduce((a, r) => a + (r.processed_count || 0), 0) /
                runs.length
            )
          : 0;
      const avgCompleted =
        runs && runs.length
          ? Math.round(
              runs.reduce((a, r) => a + (r.completed_count || 0), 0) /
                runs.length
            )
          : 0;
      const response = {
        ok: true,
        stale,
        lastRun: last,
        recentRuns: runs || [],
        stats: {
          window: runs?.length || 0,
          successes,
          failures,
          successRate:
            runs && runs.length ? +(successes / runs.length).toFixed(3) : 0,
          avgProcessed,
          avgCompleted,
        },
      };
      res.status(200).json(response);
    } catch (e: any) {
      next(
        createHttpError(
          500,
          "JOB_HEALTH_INTERNAL_ERROR",
          "Failed to compute job health",
          { error: e.message }
        )
      );
    }
  }
);

// Job runs listing (paged) for admin observability
adminRouter.get(
  "/jobs/daily-investments/runs",
  requireAdminRole,
  async (req, res) => {
    const page = parseInt((req.query.page as string) || "1", 10);
    const limit = parseInt((req.query.limit as string) || "20", 10);
    const offset = (page - 1) * limit;
    try {
      let query = supabase
        .from("job_runs")
        .select("*", { count: "exact" })
        .eq("job_name", "daily-investments")
        .order("started_at", { ascending: false })
        .range(offset, offset + limit - 1);
      const { data, error, count } = await query;
      if (error) {
        return res.status(500).json({ ok: false, error: error.message });
      }
      return res.status(200).json({
        ok: true,
        runs: data || [],
        page,
        total: count || 0,
        totalPages: Math.ceil((count || 0) / limit),
      });
    } catch (e: any) {
      return res.status(500).json({ ok: false, error: e.message });
    }
  }
);

// Transactions endpoint (used for both deposits and withdrawals)
adminRouter.get("/transactions", requireAdminRole, async (req, res) => {
  try {
    // Get query parameters
    const type = req.query.type as string;
    const status = req.query.status as string;
    const search = req.query.search as string;
    const page = parseInt(req.query.page as string) || 1;
    const limit = parseInt(req.query.limit as string) || 10;
    const offset = (page - 1) * limit;

    // Build query
    let query = supabase
      .from("transactions")
      .select(
        `
        *,
        users (
          id,
          username,
          email,
          firstName,
          lastName
        )
      `,
        { count: "exact" }
      )
      .order("created_at", { ascending: false });

    // Apply filters
    if (type) {
      query = query.eq("type", type);
    }
    if (status) {
      query = query.eq("status", status);
    }
    if (search) {
      query = query.or(
        `description.ilike.%${search}%,users.email.ilike.%${search}%`
      );
    }

    // Apply pagination
    query = query.range(offset, offset + limit - 1);

    // Execute query
    const { data: transactions, error, count } = await query;

    if (error) {
      console.error("Supabase error:", error);
      return res.status(500).json({ message: "Failed to fetch transactions" });
    }

    return res.status(200).json({
      transactions: transactions || [],
      total: count || 0,
      page,
      totalPages: Math.ceil((count || 0) / limit),
    });
  } catch (error) {
    console.error("Admin transactions fetch error:", error);
    return res.status(500).json({ message: "Failed to fetch transactions" });
  }
});

// Stats endpoint
adminRouter.get("/stats", requireAdminRole, async (req, res) => {
  try {
    // Get user counts
    const { count: totalUsers } = await supabase
      .from("users")
      .select("*", { count: "exact", head: true });

    const { count: activeUsers } = await supabase
      .from("users")
      .select("*", { count: "exact", head: true })
      .eq("is_active", true);

    // Get transaction counts and totals
    const { data: transactionStats } = await supabase
      .from("transactions")
      .select("type, status, amount, created_at");

    const stats = transactionStats?.reduce(
      (acc: any, tx) => {
        const amount = parseFloat(tx.amount) || 0;
        if (tx.type === "deposit") {
          acc.totalDeposits += amount;
          if (tx.status === "pending") acc.deposits.pending++;
          if (tx.status === "completed") acc.deposits.approved++;
          acc.deposits.total++;

          // Check if transaction is from current month
          const txDate = new Date(tx.created_at);
          const now = new Date();
          if (
            txDate.getMonth() === now.getMonth() &&
            txDate.getFullYear() === now.getFullYear()
          ) {
            acc.deposits.thisMonth += amount;
          }
        } else if (tx.type === "withdrawal") {
          acc.totalWithdrawals += amount;
          if (tx.status === "pending") acc.withdrawals.pending++;
          if (tx.status === "completed") acc.withdrawals.approved++;
          acc.withdrawals.total++;

          // Check if transaction is from current month
          const txDate = new Date(tx.created_at);
          const now = new Date();
          if (
            txDate.getMonth() === now.getMonth() &&
            txDate.getFullYear() === now.getFullYear()
          ) {
            acc.withdrawals.thisMonth += amount;
          }
        }
        if (tx.status === "pending") acc.pendingTransactions++;
        return acc;
      },
      {
        totalDeposits: 0,
        totalWithdrawals: 0,
        pendingTransactions: 0,
        deposits: {
          total: 0,
          pending: 0,
          approved: 0,
          thisMonth: 0,
        },
        withdrawals: {
          total: 0,
          pending: 0,
          approved: 0,
          thisMonth: 0,
        },
      }
    ) || {
      totalDeposits: 0,
      totalWithdrawals: 0,
      pendingTransactions: 0,
      deposits: { total: 0, pending: 0, approved: 0, thisMonth: 0 },
      withdrawals: { total: 0, pending: 0, approved: 0, thisMonth: 0 },
    };

    return res.status(200).json({
      totalUsers: totalUsers || 0,
      activeUsers: activeUsers || 0,
      ...stats,
      maintenanceMode: false,
    });
  } catch (error) {
    console.error("Admin stats fetch error:", error);
    return res.status(500).json({ message: "Failed to fetch admin stats" });
  }
});

// Audit logs endpoint
adminRouter.get("/audit-logs", requireAdminRole, async (req, res) => {
  try {
    // Get query parameters
    const page = parseInt(req.query.page as string) || 1;
    const limit = parseInt(req.query.limit as string) || 10;
    const offset = (page - 1) * limit;
    const action = req.query.action as string;
    const search = req.query.search as string;

    // Build query
    let query = supabase
      .from("audit_logs")
      .select(
        `
        *,
        users (
          id,
          username,
          email
        )
      `,
        { count: "exact" }
      )
      .order("created_at", { ascending: false });

    // Apply filters
    if (action) {
      query = query.eq("action", action);
    }
    if (search) {
      query = query.or(
        `description.ilike.%${search}%,action.ilike.%${search}%`
      );
    }

    // Apply pagination
    query = query.range(offset, offset + limit - 1);

    // Execute query
    const { data: logs, error, count } = await query;

    if (error) {
      console.error("Supabase error:", error);
      return res.status(500).json({ message: "Failed to fetch audit logs" });
    }

    return res.status(200).json({
      logs: logs || [],
      totalLogs: count || 0,
      page,
      totalPages: Math.ceil((count || 0) / limit),
    });
  } catch (error) {
    console.error("Audit logs fetch error:", error);
    return res.status(500).json({ message: "Failed to fetch audit logs" });
  }
});

// Update transaction status
adminRouter.patch(
  "/transactions/:id/status",
  requireAdminRole,
  async (req, res) => {
    try {
      const { id } = req.params;
      const { status } = req.body;

      // First get the transaction details
      const { data: transaction, error: fetchError } = await supabase
        .from("transactions")
        .select(
          `
        *,
        users (
          id,
          username,
          email,
          firstName,
          lastName
        )
      `
        )
        .eq("id", id)
        .single();

      if (fetchError || !transaction) {
        console.error("Failed to fetch transaction:", fetchError);
        return res.status(404).json({ message: "Transaction not found" });
      }

      // Update the transaction status
      const { data, error } = await supabase
        .from("transactions")
        .update({ status })
        .eq("id", id)
        .select()
        .single();

      if (error) {
        console.error("Supabase error:", error);
        return res
          .status(500)
          .json({ message: "Failed to update transaction status" });
      }

      // Track whether an approval email was sent (surface to client for observability)
      let emailSent: boolean | undefined = undefined;
      let emailOutcome: string | undefined = undefined;

      // Send email notifications when transactions are approved
      if (status === "completed") {
        const user = transaction.users;
        const amount = transaction.amount;
        try {
          if (transaction.type === "deposit") {
            try {
              const primaryOk = await sendDepositApprovedEmail(
                user,
                amount,
                transaction.description || "Investment Deposit"
              );
              emailSent = primaryOk;
              emailOutcome = primaryOk ? "primary-success" : "primary-failed";
              if (!primaryOk) {
                try {
                  const { sendDepositApprovedEmail: mgrSend } = await import(
                    "./emailManager"
                  );
                  const fallbackOk = await mgrSend(
                    user,
                    amount,
                    transaction.description || "Investment Deposit"
                  );
                  emailSent = fallbackOk;
                  emailOutcome = fallbackOk
                    ? "fallback-success"
                    : "fallback-failed";
                } catch (mgrErr) {
                  console.warn(
                    "[email-debug] legacy deposit approval fallback threw",
                    mgrErr
                  );
                  emailOutcome = "fallback-exception";
                }
              }
            } catch (primaryErr) {
              console.warn(
                "[email-debug] legacy deposit approval primary threw",
                primaryErr
              );
              emailSent = false;
              emailOutcome = "primary-exception";
            }
          } else if (transaction.type === "withdrawal") {
            try {
              const primaryOk = await sendWithdrawalApprovedEmail(
                user,
                amount,
                transaction.destination || "Your Crypto Wallet"
              );
              emailSent = primaryOk;
              emailOutcome = primaryOk ? "primary-success" : "primary-failed";
              if (!primaryOk) {
                try {
                  const { sendWithdrawalApprovedEmail: mgrSend } = await import(
                    "./emailManager"
                  );
                  const fallbackOk = await mgrSend(
                    user,
                    amount,
                    transaction.destination || "Your Crypto Wallet"
                  );
                  emailSent = fallbackOk;
                  emailOutcome = fallbackOk
                    ? "fallback-success"
                    : "fallback-failed";
                } catch (mgrErr) {
                  console.warn(
                    "[email-debug] legacy withdrawal approval fallback threw",
                    mgrErr
                  );
                  emailOutcome = "fallback-exception";
                }
              }
            } catch (primaryErr) {
              console.warn(
                "[email-debug] legacy withdrawal approval primary threw",
                primaryErr
              );
              emailSent = false;
              emailOutcome = "primary-exception";
            }
          }
        } catch (outerErr) {
          console.error("Failed to send notification email:", outerErr);
          emailSent = false;
          if (!emailOutcome) emailOutcome = "exception";
        }
      }
      // Attach emailSent + outcome if defined for frontend/admin UI diagnostics
      const responsePayload =
        emailSent === undefined ? data : { ...data, emailSent, emailOutcome };
      return res.status(200).json(responsePayload);
    } catch (error) {
      console.error("Update transaction status error:", error);
      return res
        .status(500)
        .json({ message: "Failed to update transaction status" });
    }
  }
);

// Bulk update transaction status
adminRouter.post(
  "/transactions/bulk-status",
  requireAdminRole,
  async (req, res) => {
    try {
      const { ids, status } = req.body;

      const { data, error } = await supabase
        .from("transactions")
        .update({ status })
        .in("id", ids)
        .select();

      if (error) {
        console.error("Supabase error:", error);
        return res
          .status(500)
          .json({ message: "Failed to update transactions" });
      }

      return res.status(200).json(data);
    } catch (error) {
      console.error("Bulk update transactions error:", error);
      return res.status(500).json({ message: "Failed to update transactions" });
    }
  }
);

// User Management Endpoints

// Get all users (simplified list)
adminRouter.get("/users/list", requireAdminRole, async (req, res) => {
  try {
    const { data: users, error } = await supabase
      .from("users")
      .select(
        "id, username, email, first_name, last_name, role, created_at, email_confirmed_at"
      )
      .order("created_at", { ascending: false });

    if (error) {
      console.error("Failed to fetch users:", error);
      return res.status(500).json({ message: "Failed to fetch users" });
    }

    return res.status(200).json({ users: users || [] });
  } catch (error) {
    console.error("Get users error:", error);
    return res.status(500).json({ message: "Failed to fetch users" });
  }
});

// Delete user entirely
adminRouter.delete(
  "/users/:userId/delete",
  requireAdminRole,
  async (req, res) => {
    try {
      const userId = parseInt(req.params.userId);

      if (isNaN(userId)) {
        return res.status(400).json({ message: "Invalid user ID" });
      }

      // First, get user info for logging
      const { data: user, error: getUserError } = await supabase
        .from("users")
        .select("email, username, role")
        .eq("id", userId)
        .single();

      if (getUserError || !user) {
        return res.status(404).json({ message: "User not found" });
      }

      // Prevent deletion of admin users (optional safety check)
      if (user.role === "admin") {
        return res.status(403).json({
          message: "Cannot delete admin users. Please change role first.",
        });
      }

      // Delete related data first to avoid foreign key constraints
      // Delete transactions
      await supabase.from("transactions").delete().eq("user_id", userId);

      // Delete user
      const { error: deleteError } = await supabase
        .from("users")
        .delete()
        .eq("id", userId);

      if (deleteError) {
        console.error("Failed to delete user:", deleteError);
        return res.status(500).json({ message: "Failed to delete user" });
      }

      console.log(`✅ User deleted: ${user.email} (ID: ${userId})`);

      return res.status(200).json({
        message: "User deleted successfully",
        deletedUser: {
          id: userId,
          email: user.email,
          username: user.username,
        },
      });
    } catch (error) {
      console.error("Delete user error:", error);
      return res.status(500).json({ message: "Failed to delete user" });
    }
  }
);

// Change user password
adminRouter.post(
  "/users/:userId/change-password",
  requireAdminRole,
  async (req, res) => {
    try {
      const userId = parseInt(req.params.userId);
      const { newPassword } = req.body;

      if (isNaN(userId)) {
        return res.status(400).json({ message: "Invalid user ID" });
      }

      if (!newPassword || newPassword.length < 6) {
        return res.status(400).json({
          message: "Password must be at least 6 characters long",
        });
      }

      // Get user info
      const { data: user, error: getUserError } = await supabase
        .from("users")
        .select("email, username")
        .eq("id", userId)
        .single();

      if (getUserError || !user) {
        return res.status(404).json({ message: "User not found" });
      }

      // Update password using Supabase Admin API
      const { error: updateError } = await supabase.auth.admin.updateUserById(
        userId.toString(),
        {
          password: newPassword,
        }
      );

      if (updateError) {
        console.error("Failed to update password:", updateError);
        return res.status(500).json({
          message: "Failed to update password: " + updateError.message,
        });
      }

      console.log(
        `✅ Password changed for user: ${user.email} (ID: ${userId})`
      );

      return res.status(200).json({
        message: "Password updated successfully",
        user: {
          id: userId,
          email: user.email,
          username: user.username,
        },
      });
    } catch (error) {
      console.error("Change password error:", error);
      return res.status(500).json({ message: "Failed to change password" });
    }
  }
);

export default adminRouter;